Move integration tests under 'integration' package
