+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/integration/SentryPluginWithFirebaseTest.kt
@@ -1,4 +1,4 @@
-package io.sentry.android.gradle
+package io.sentry.android.gradle.integration
 
 import kotlin.test.assertTrue
 import org.junit.Test
